+++ CHANGELOG.md
@@ -8,7 +8,8 @@ To install or update LODA, please follow the [installation instructions](https:/
 
 ### Features
 
-* Support comparision operators in formula generation
+* Support comparison operators in formula generation
+* Support `dgr` operation in formula generation
 * Support factorial in formula generation
 
 ### Enhancements